Improve LDraw Visualiser script execution and error reporting
- Enhance NullProxy with implicit conversions for math types to allow scripts to run without an active debugger.
- Improve the error panel UI using Visual Studio theme colors, better text wrapping, and more descriptive diagnostic messages.
- Add extension icons and preview images to the VSIX manifest.
- Update NuGet references to use wildcard versions and simplified LDraw builder command handling.

+++ projects/rylogic/Rylogic.Gfx/src/LDraw/Builder.cs
@@ -1275,7 +1275,7 @@ namespace Rylogic.LDraw
 				{
 					res.Write(EKeyword.Data, () =>
 					{
-						res.Append((int)cmd.m_id);
+						res.Append(cmd.m_id);
 						foreach (var p in cmd.m_params)
 						{
 							if (p is bool bool_) { res.Append(bool_); continue; }
